Health station (feeding(part2))

+++ src/main/java/lv/rvt/Person.java
@@ -65,7 +65,7 @@ public class Person {
         this.height = newHeight;
     }
 
-    public void setWeight(int newWeight) {
+    public void setWeight(double newWeight) {
         this.weight = newWeight;
     }
